% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coretools-package.R
\docType{package}
\name{coretools-package}
\alias{coretools}
\alias{coretools-package}
\title{A Pack of Help Functions Used in the Core}
\description{
A Pack of Help Functions Used in the Core
}
\keyword{package}

