% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_core.R
\name{install_core_package}
\alias{install_core_package}
\title{Install core package}
\usage{
install_core_package(option = c("normal", "minimum", "full"), bioc = FALSE)
}
\arguments{
\item{option}{Choose among \code{normal}, \code{minimum} and
\code{full} to install different packages under different cases}

\item{bioc}{If anyone want to install packages from bioconductor, set
this option to be \code{TRUE}.}
}
\description{
Install a pack of R packages that people usually use around the
core. Anyone are welcome to contribute
}

