% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_initiate.R
\name{project_initiate}
\alias{project_initiate}
\title{Bootstrap an analytical folder}
\usage{
project_initiate(targetdir = ".")
}
\arguments{
\item{targetdir}{The default setting is to initial a new project folder
in the current working directory. However, if you want to create this
skeleton in a subdirectory, use this option}
}
\description{
Bootstrap an analytical folder
}

